+++ src/Symfony/Component/Serializer/CHANGELOG.md
@@ -7,7 +7,7 @@ CHANGELOG
  * added the list of constraint violations' parameters in `ConstraintViolationListNormalizer`
  * added support for serializing `DateTimeZone` objects
  * added a `deep_object_to_populate` context option to recursive denormalize on `object_to_populate` object.
- * added `ScalarDenormalizer` for denormalizing scalar values
+ * added support for denormalizing scalar values
 
 4.2.0
 -----